requirements.txt
flask
psycopg2-binary
gunicorn
azure-identity